removed faulty func

+++ android/src/main/java/com/swmansion/rnscreens/events/HeaderHeightChangeEvent.kt
@@ -7,16 +7,16 @@ import com.facebook.react.uimanager.events.Event
 class HeaderHeightChangeEvent(
     surfaceId: Int,
     viewId: Int,
-    private val headerHeight: Double,
+    private val headerHeight: Int,
 ) : Event<HeaderHeightChangeEvent>(surfaceId, viewId) {
     override fun getEventName() = EVENT_NAME
 
     // As the same header height could appear twice, use header height as a coalescing key.
-    override fun getCoalescingKey(): Short = headerHeight.toInt().toShort()
+    override fun getCoalescingKey(): Short = headerHeight.toShort()
 
     override fun getEventData(): WritableMap? =
         Arguments.createMap().apply {
-            putDouble("headerHeight", headerHeight)
+            putDouble("headerHeight", headerHeight.toDouble())
         }
 
     companion object {
